Add scripts, add variable for display names
