+++ .gitignore
@@ -14,3 +14,5 @@ terraform.tfvars
 # jetbrains
 .idea
 /test/
+/scripts/spot.egg-info/
+scripts/__pycache__/
