+++ examples/main.tf
@@ -8,8 +8,12 @@ provider "azuread" {
 
 #Call the spot module to create a Spot account and link Azure
 module "azure-connect" {
-  source = "spotinst/azure-connect/spotinst"
+  source = "../"
+
+  spotinst_token = "redacted"
 }
+
+
 output "spot_account_id" {
   value = module.azure-connect.spot_account_id
 }
