+++ scripts/spot-account
@@ -3,4 +3,4 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 cd "${DIR}"
 pip3 install -e . > /dev/null
-spot-account-azure "$@"
+spot-account "$@"
