--- scripts/spot_account_azure.egg-info/PKG-INFO
@@ -1,10 +1,0 @@
-Metadata-Version: 2.1
-Name: spot-account-azure
-Version: 0.1
-Summary: UNKNOWN
-Home-page: UNKNOWN
-License: UNKNOWN
-Platform: UNKNOWN
-
-UNKNOWN
-
