--- scripts/spot_account_azure.egg-info/dependency_links.txt
@@ -1,1 +1,0 @@
-
