--- scripts/spot_account_azure.egg-info/entry_points.txt
@@ -1,3 +1,0 @@
-[console_scripts]
-spot-account-azure = spot_account_azure:cli
-
