--- scripts/spot_account_azure.egg-info/requires.txt
@@ -1,2 +1,0 @@
-Click
-spotinst-sdk2>=2.1.10
