--- scripts/spot_account_azure.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-
